Rework primitives

+++ command_queue.go
@@ -101,7 +101,7 @@ const (
 	// Note: The reference count returned should be considered immediately stale. It is unsuitable for
 	// general use in applications. This feature is provided for identifying memory leaks.
 	//
-	// Returned type: Uint
+	// Returned type: uint32
 	QueueReferenceCountInfo CommandQueueInfoName = C.CL_QUEUE_REFERENCE_COUNT
 	// QueuePropertiesInfo returns the currently specified properties for the command-queue.
 	// These properties are specified by the value associated with the QueuePropertiesProperty passed in as
@@ -122,7 +122,7 @@ const (
 // Raw strings are with a terminating NUL character. For convenience, use CommandQueueInfoString().
 //
 // See also: https://registry.khronos.org/OpenCL/sdk/1.2/docs/man/xhtml/clGetCommandQueueInfo.html
-func CommandQueueInfo(commandQueue CommandQueue, paramName CommandQueueInfoName, paramSize uint, paramValue unsafe.Pointer) (uint, error) {
+func CommandQueueInfo(commandQueue CommandQueue, paramName CommandQueueInfoName, paramSize uintptr, paramValue unsafe.Pointer) (uintptr, error) {
 	sizeReturn := C.size_t(0)
 	status := C.clGetCommandQueueInfo(
 		commandQueue.handle(),
@@ -133,7 +133,7 @@ func CommandQueueInfo(commandQueue CommandQueue, paramName CommandQueueInfoName,
 	if status != C.CL_SUCCESS {
 		return 0, StatusError(status)
 	}
-	return uint(sizeReturn), nil
+	return uintptr(sizeReturn), nil
 }
 
 // Flush issues all previously queued OpenCL commands in a command-queue to the device associated with the
